mail list and credentials from db and userdao fix findById int

+++ web/src/main/java/com/infoshareacademy/jbusters/dao/UserDao.java
@@ -35,7 +35,7 @@ public class UserDao {
         return entityManager.find(User.class, userEmail);
     }
 
-    public User findById(Long id) {
+    public User findById(int id) {
         return entityManager.find(User.class, id);
     }
